place anonymize_user in user model file
